Rename followup and payment mode pages
