+++ src/App.jsx
@@ -12,7 +12,7 @@ import Signup from './components/Signup';
 import OrgCategories from './pages/OrgCategories';
 import Education from './pages/Education';
 import Exam from './pages/Exam';
-import PaymentMode from './pages/remove1';
+import PaymentMode from './pages/PaymentMode';
 import ForgotPassword from './components/ForgotPassword';
 import ResetPassword from './components/ResetPassword';
 import InstituteProfile from './pages/instituteProfile';
@@ -23,7 +23,7 @@ import Leads from './reports/Leads';
 import AllAdmission from './reports/allAdmission';
 import AddLead from './pages/AddLead';
 import AddNew from './components/admissions/AddAdmission';
-import Followup from './pages/remove';
+import Followup from './pages/Followup';
 import WhatsAppAdminPage from './pages/WhatsAppAdminPage';
 import AddReciept from './pages/addReciept';
 import AddPayment from './pages/addPayment';
